Add transaction service and related exceptions for currency operations (using FXRatesAPI)

+++ src/test/java/lu/llemaire/xchange_transfer/services/AccountTransactionalServiceTest.java
@@ -5,17 +5,19 @@ import lu.llemaire.xchange_transfer.account.AccountRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
+
 @Service
 @Transactional
 @RequiredArgsConstructor
 public class AccountTransactionalServiceTest {
     final AccountRepository accountRepository;
 
-    public int addBalance(Long accountId, Double amount) {
+    public int addBalance(Long accountId, BigDecimal amount) {
         return accountRepository.addBalance(accountId, amount);
     }
 
-    public int subtractBalance(Long accountId, Double amount) {
+    public int subtractBalance(Long accountId, BigDecimal amount) {
         return accountRepository.subtractBalance(accountId, amount);
     }
 }
